requirements.txt
fastapi==0.115.0
perspective-python==3.1.1
numpy==1.26.4
pyarrow==16.1.0
websocket-client==1.8.0
databricks-sql-connector==3.2.0
azure-monitor-opentelemetry==1.6.0
fastapi-cache2==0.2.1
opentelemetry.instrumentation.fastapi==0.46b0
kafka-python==2.0.2
avro==1.11.3
fastapi_health
python-snappy==0.7.2
uvicorn[standard]
pydantic==2.9.2
pydantic-settings
python-dotenv
loguru
